Reorganized Routers
* Moved endpoints to appropriately named router files to clean up code
* Renamed some router variables and added some typing hints
